[refactor] Merge b1z1_config and manip_loco_config into a single file for readablility
