+++ low-level/legged_gym/envs/manip_loco/b1z1_config.py
@@ -28,27 +28,51 @@
 #
 # Copyright (c) 2021 ETH Zurich, Nikita Rudin
 
-from .manip_loco_config import ManipLocoCfg, ManipLocoCfgPPO
+from legged_gym.envs.base.legged_robot_config import LeggedRobotCfg, LeggedRobotCfgPPO
 import numpy as np
 
-class B1Z1RoughCfg( ManipLocoCfg ):
-    class goal_ee ( ManipLocoCfg.goal_ee ):
+class B1Z1RoughCfg( LeggedRobotCfg ):
+    class goal_ee:
+        num_commands = 3
+        traj_time = [1, 3]
+        hold_time = [0.5, 2]
         collision_upper_limits = [0.1, 0.2, -0.05]
         collision_lower_limits = [-0.8, -0.2, -0.7]
         underground_limit = -0.7
+        num_collision_check_samples = 10
+        command_mode = 'sphere'
         arm_induced_pitch = 0.38 # Added to -pos_p (negative goal pitch) to get default eef orn_p
-    
-        class sphere_center ( ManipLocoCfg.goal_ee.sphere_center ):
+
+        class sphere_center:
             x_offset = 0.3 # Relative to base
             y_offset = 0 # Relative to base
             z_invariant_offset = 0.7 # Relative to terrain
-        
-        class ranges(ManipLocoCfg.goal_ee.ranges):
+
+        class ranges:
             init_pos_start = [0.5, np.pi/8, 0]
             init_pos_end = [0.7, 0, 0]
             pos_l = [0.4, 0.95]
             pos_p = [-1 * np.pi / 2.5, 1 * np.pi / 3]
             pos_y = [-1.2, 1.2]
+            
+            delta_orn_r = [-0.5, 0.5]
+            delta_orn_p = [-0.5, 0.5]
+            delta_orn_y = [-0.5, 0.5]
+            final_tracking_ee_reward = 0.55
+
+        sphere_error_scale = [1, 1, 1]#[1 / (ranges.final_pos_l[1] - ranges.final_pos_l[0]), 1 / (ranges.final_pos_p[1] - ranges.final_pos_p[0]), 1 / (ranges.final_pos_y[1] - ranges.final_pos_y[0])]
+        orn_error_scale = [1, 1, 1]#[2 / np.pi, 2 / np.pi, 2 / np.pi]
+
+    class noise:
+        add_noise = False
+        noise_level = 1.0 # scales other values
+        class noise_scales:
+            dof_pos = 0.01
+            dof_vel = 1.5
+            lin_vel = 0.1
+            ang_vel = 0.2
+            gravity = 0.05
+            height_measurements = 0.1   
 
     class commands:
         curriculum = True
@@ -61,28 +85,42 @@ class B1Z1RoughCfg( ManipLocoCfg ):
 
         ang_vel_yaw_clip = 0.5
         lin_vel_x_clip = 0.2
-        ang_vel_pitch_clip = ang_vel_yaw_clip
 
         class ranges:
             lin_vel_x = [-0.8, 0.8] # min max [m/s]
-            # lin_vel_y = [0, 0]   # min max [m/s]
             ang_vel_yaw = [-1.0, 1.0]    # min max [rad/s]
-            ang_vel_pitch = [-1.0, 1.0]    # min max [rad/s]
 
-    class env (ManipLocoCfg.env):
-        num_gripper_joints = 1 # 2 for two finger gripper
+    class normalization:
+        class obs_scales:
+            lin_vel = 1.0
+            ang_vel =  1.0
+            dof_pos = 1.0
+            dof_vel = 0.05
+            height_measurements = 5.0
+        clip_observations = 100.
+        clip_actions = 100.
+
+    class env:
+        num_envs = 6144
+        num_actions = 12 + 6 #CAUTION
+        num_torques = 12 + 6
+        action_delay = 3  # -1 for no delay
+        num_gripper_joints = 1
+        num_proprio = 2 + 3 + 18 + 18 + 12 + 4 + 3 + 3 + 3 
         num_priv = 5 + 1 + 12
         history_len = 10
-        num_proprio = 2 + 3 + 18 + 18 + 12 + 4 + 3 + 3 + 3
         num_observations = num_proprio * (history_len+1) + num_priv
-        action_delay = 3 # Not used, assigned in code
+        num_privileged_obs = None # if not None a priviledge_obs_buf will be returned by step() (critic obs for assymetric training). None is returned otherwise 
+        send_timeouts = True # send time out information to the algorithm
+        episode_length_s = 10 # episode length in seconds
+        reorder_dofs = True
+        teleop_mode = False # Overriden in teleop.py. When true, commands come from keyboard
+        record_video = False
+        stand_by = False
         observe_gait_commands = False
         frequencies = 2
-        observe_velocities = True
-        stand_only = False
-        pitch_control = False
 
-    class init_state( ManipLocoCfg.init_state ):
+    class init_state( LeggedRobotCfg.init_state ):
         pos = [0.0, 0.0, 0.5] # x,y,z [m]
         default_joint_angles = { # = target angles [rad] when action = 0.0
             'FL_hip_joint': 0.2,   # [rad]
@@ -103,59 +141,128 @@ class B1Z1RoughCfg( ManipLocoCfg ):
 
             'z1_waist': 0.0,
             'z1_shoulder': 1.48,
-            'z1_elbow': -1.5, # -0.63,
-            'z1_wrist_angle': 0, # -0.84,
+            'z1_elbow': -0.63,
+            'z1_wrist_angle': -0.84,
             'z1_forearm_roll': 0.0,
-            'z1_wrist_rotate': 1.57, # 0.0,
+            'z1_wrist_rotate': 1.57,#0.0,
             'z1_jointGripper': -0.785,
         }
+        rand_yaw_range = np.pi/2
+        origin_perturb_range = 0.5
+        init_vel_perturb_range = 0.1
 
-    class control( ManipLocoCfg.control ):
+    class control:
         stiffness = {'joint': 80, 'z1': 5}  # [N*m/rad] # Kp: 80, 150, 200
         damping = {'joint': 2.0, 'z1': 0.5}     # [N*m*s/rad]
 
-    class asset( ManipLocoCfg.asset ):
+        adaptive_arm_gains = False
+        # action scale: target angle = actionScale * action + defaultAngle
+        action_scale = [0.4, 0.45, 0.45] * 2 + [0.4, 0.45, 0.45] * 2 + [2.1, 0.6, 0.6, 0, 0, 0]
+        # decimation: Number of control action updates @ sim DT per policy DT
+        decimation = 4
+        torque_supervision = False
+
+    class asset( LeggedRobotCfg.asset ):
         file = '{LEGGED_GYM_ROOT_DIR}/resources/robots/b1z1/urdf/b1z1.urdf'
         foot_name = "foot"
         gripper_name = "ee_gripper_link" #"gripperMover"
-        # gripper_name = "wx250s/ee_gripper_link" # for two finger gripper
         penalize_contacts_on = ["thigh", "trunk", "calf"]
         terminate_after_contacts_on = []
+        self_collisions = 0 # 1 to disable, 0 to enable...bitwise filter
+        flip_visual_attachments = False
+        collapse_fixed_joints = True # Specific fixed joints can be kept by adding " <... dont_collapse="true">
+        fix_base_link = False
+    
+    class box:
+        box_size = 0.1
+        randomize_base_mass = True
+        added_mass_range = [-0.001, 0.050]
+        box_env_origins_x = 0
+        box_env_origins_y_range = [0.1, 0.3]
+        box_env_origins_z = box_size / 2 + 0.16
+    
+    class arm:
+        init_target_ee_base = [0.2, 0.0, 0.2]
+        grasp_offset = 0.08
+        osc_kp = np.array([100, 100, 100, 30, 30, 30])
+        osc_kd = 2 * (osc_kp ** 0.5)
+
+    class domain_rand:
+        observe_priv = True
+        randomize_friction = True
+        friction_range = [0.3, 3.0] # [0.5, 3.0]
+        randomize_base_mass = True
+        added_mass_range = [0., 15.]
+        randomize_base_com = True
+        added_com_range_x = [-0.15, 0.15]
+        added_com_range_y = [-0.15, 0.15]
+        added_com_range_z = [-0.15, 0.15]
+        randomize_motor = True
+        leg_motor_strength_range = [0.7, 1.3]
+        arm_motor_strength_range = [0.7, 1.3]
+        randomize_gripper_mass = True
+        gripper_added_mass_range = [0, 0.1]
+        push_robots = True
+        push_interval_s = 8
+        max_push_vel_xy = 0.5
   
-    class rewards ( ManipLocoCfg.rewards ):
+    class rewards:
+        reward_container_name = "maniploco_rewards"   # select the reward container to use
+
+        # -------Common Para. ---------
+        only_positive_rewards = False # if true negative total rewards are clipped at zero (avoids early termination problems)
+        tracking_sigma = 0.2  # tracking reward = exp(-error^2/sigma)
+        tracking_ee_sigma = 1
+        soft_dof_pos_limit = 1.  # percentage of urdf limits, values above this limit are penalized
+        soft_dof_vel_limit = 1.
+        soft_torque_limit = 0.4
+        base_height_target = 0.55
+        max_contact_force = 40.  # forces above this value are penalized
+        # -------Gait control Para. ---------
         gait_vel_sigma = 0.5
         gait_force_sigma = 0.5
         kappa_gait_probs = 0.07
-        class scales ( ManipLocoCfg.rewards.scales ):
+        feet_height_target = 0.3
+
+        feet_aritime_allfeet = False
+        feet_height_allfeet = False
+
+        # Scales set to 0 will still be logged (as zero reward and non-zero metric)
+        # To not compute and log a given metric, set the scale to None
+        class scales:
+            # -------Gait control rewards ---------
             tracking_contacts_shaped_force = -2.0 # Only works when `observing_gait_commands` is true
             tracking_contacts_shaped_vel = -2.0 # Only works when `observing_gait_commands` is true
-            tracking_lin_vel_max = 2.0 # 1.5
+            feet_air_time = 2.0
+            feet_height = 1.0
+
+            # -------Tracking rewards ----------
+            tracking_lin_vel_max = 2.0 
             tracking_lin_vel_x_l1 = 0.
             tracking_lin_vel_x_exp = 0
-            tracking_ang_vel = 0.5 # just for yaw
+            tracking_ang_vel = 0.5
+
             delta_torques = -1.0e-7/4.0
             work = 0
             energy_square = 0.0
-            torques = -2.5e-5 # -1e-5
-            stand_still = 1.0 #1.5
+            torques = -2.5e-5 
+            stand_still = 1.0 
             walking_dof = 1.5
             dof_default_pos = 0.0
-            dof_error = 0.0 # -0.06 # -0.04
+            dof_error = 0.0 
             alive = 1.0
             lin_vel_z = -1.5
-            roll = -2.0
-            
-            # tracking_ang_pitch_vel = 0.5 # New reward, only useful when pitch_control = True
+            roll = -2
 
             # common rewards
-            feet_air_time = 1.0
-            feet_height = 1.0
-            ang_vel_xy = -0.2 # -0.1
-            dof_acc = -7.5e-7 #-2.5e-7
+            ang_vel_xy = -0.2 
+            dof_acc = -7.5e-7 
             collision = -10.
             action_rate = -0.015
             dof_pos_limits = -10.0
+            delta_torques = -1.0e-7
             hip_pos = -0.3
+            work = -0.003
             feet_jerk = -0.0002
             feet_drag = -0.08
             feet_contact_forces = -0.001
@@ -165,12 +272,13 @@ class B1Z1RoughCfg( ManipLocoCfg ):
             base_height = -5.0
             torques_walking = 0.0
             torques_standing = 0.0
+            energy_square = 0.0
             energy_square_walking = 0.0
             energy_square_standing = 0.0
             base_height_walking = 0.0
             base_height_standing = 0.0
-            penalty_lin_vel_y = 0.#-10.
-        base_height_target = 0.55
+            penalty_lin_vel_y = 0.
+
         class arm_scales:
             arm_termination = None
             tracking_ee_sphere = 0.
@@ -182,15 +290,122 @@ class B1Z1RoughCfg( ManipLocoCfg ):
             arm_energy_abs_sum = None
             tracking_ee_orn = 0.
             tracking_ee_orn_ry = None
+        
+    class viewer:
+        pos = [-20, 0, 20]  # [m]
+        lookat = [0, 0, -2]  # [m]
 
-class B1Z1RoughCfgPPO(ManipLocoCfgPPO):
-    class policy (ManipLocoCfgPPO.policy):
+    class termination:
+        r_threshold = 0.8
+        p_threshold = 0.8
+        z_threshold = 0.1
+
+    class terrain:
+        mesh_type = 'trimesh' # "heightfield" # none, plane, heightfield or trimesh
+        hf2mesh_method = "fast"  # grid or fast
+        max_error = 0.1 # for fast
+        horizontal_scale = 0.05 # [m] influence computation time by a lot
+        vertical_scale = 0.005 # [m]
+        border_size = 25 # [m]
+        height = [0.00, 0.1] # [0.04, 0.1]
+        gap_size = [0.02, 0.1]
+        stepping_stone_distance = [0.02, 0.08]
+        downsampled_scale = 0.075
+        curriculum = False
+
+        all_vertical = False
+        no_flat = True
+        
+        static_friction = 1.0
+        dynamic_friction = 1.0
+        restitution = 0.
+
+        measure_heights = True
+        measured_points_x = [-0.8, -0.7, -0.6, -0.5, -0.4, -0.3, -0.2, -0.1, 0., 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8] # 1mx1.6m rectangle (without center line)
+        measured_points_y = [-0.5, -0.4, -0.3, -0.2, -0.1, 0., 0.1, 0.2, 0.3, 0.4, 0.5]
+        
+        selected = False # select a unique terrain type and pass all arguments
+        terrain_kwargs = None # Dict of arguments for selected terrain
+        max_init_terrain_level = 5 # starting curriculum state
+        terrain_length = 8.
+        terrain_width = 8.
+        num_rows= 10 # number of terrain rows (levels)  # spreaded is benifitiall !
+        num_cols = 20 # number of terrain cols (types)
+
+        terrain_dict = {"smooth slope": 0., 
+                        "rough slope up": 0.,
+                        "rough slope down": 0.,
+                        "rough stairs up": 0., 
+                        "rough stairs down": 0., 
+                        "discrete": 0., 
+                        "stepping stones": 0.,
+                        "gaps": 0., 
+                        "rough flat": 1.0,
+                        "pit": 0.0,
+                        "wall": 0.0}
+        terrain_proportions = list(terrain_dict.values())
+        # trimesh only:
+        slope_treshold = None # slopes above this threshold will be corrected to vertical surfaces
+        origin_zero_z = False
+
+
+class B1Z1RoughCfgPPO(LeggedRobotCfgPPO):
+    seed = 1
+    runner_class_name = 'OnPolicyRunner'
+    class policy:
+        continue_from_last_std = True
+        init_std = [[0.8, 1.0, 1.0] * 4 + [1.0] * 6]
+        actor_hidden_dims = [128]
+        critic_hidden_dims = [128]
+        activation = 'elu' # can be elu, relu, selu, crelu, lrelu, tanh, sigmoid
+        output_tanh = False
+
+        leg_control_head_hidden_dims = [128, 128]
+        arm_control_head_hidden_dims = [128, 128]
+
+        priv_encoder_dims = [64, 20]
+
+        num_leg_actions = 12
+        num_arm_actions = 6
+
         adaptive_arm_gains = B1Z1RoughCfg.control.adaptive_arm_gains
+        adaptive_arm_gains_scale = 10.0
+        
+    class algorithm:
+        # training params
+        value_loss_coef = 1.0
+        use_clipped_value_loss = True
+        clip_param = 0.2
+        entropy_coef = 0.0
+        num_learning_epochs = 5
+        num_mini_batches = 4 # mini batch size = num_envs*nsteps / nminibatches
+        learning_rate = 2e-4 
+        schedule = 'fixed' # could be adaptive, fixed
+        gamma = 0.99
+        lam = 0.95
+        desired_kl = None
+        max_grad_norm = 1.
+        min_policy_std = [[0.15, 0.25, 0.25] * 4 + [0.2] * 3 + [0.05] * 3]
 
-    class algorithm (ManipLocoCfgPPO.algorithm):
+        mixing_schedule=[1.0, 0, 3000] #if not RESUME else [1.0, 0, 1]
         torque_supervision = B1Z1RoughCfg.control.torque_supervision  #alert: also appears above
+        torque_supervision_schedule=[0.0, 1000, 1000]
         adaptive_arm_gains = B1Z1RoughCfg.control.adaptive_arm_gains
+        # dagger params
+        dagger_update_freq = 20
+        priv_reg_coef_schedual = [0, 0.1, 3000, 7000] #if not RESUME else [0, 1, 1000, 1000]
 
-    class runner (ManipLocoCfgPPO.runner):
-        max_iterations = 40000
+    class runner:
+        policy_class_name = 'ActorCritic'
+        algorithm_class_name = 'PPO'
+        num_steps_per_env = 24
+        max_iterations = 45000 # number of policy updates
+        # logging
+        save_interval = 200 # check for potential saves every this many iterations
         experiment_name = 'b1z1_v2'
+        run_name = ''
+        # load and resume
+        resume = False
+        load_run = -1 # -1 = last run
+        checkpoint = -1 # -1 = last saved model
+        resume_path = None # updated from load_run and chkpt
